Add new SQL models for OCS GE artificial differences and commune intersections
- Created `artif_difference.sql` to calculate differences in OCS GE data, including geometry and attributes related to artificiality.
- Introduced `artif_difference_commune.sql` to segment OCS GE differences by commune, preserving total surface area for intersected geometries.
- Updated `schema.yml` to include data tests for the new `ocsge_artif_difference_staging` model, ensuring data integrity and relationships.
