+++ airflow/include/data/ocsge/normalization/__init__.py
@@ -1,0 +1,19 @@
+from .get_normalized_fields import get_normalized_fields
+from .ocsge_artif_diff_normalization_sql import ocsge_artif_diff_normalization_sql
+from .ocsge_artif_normalization_sql import ocsge_artif_normalization_sql
+from .ocsge_diff_normalization_sql import ocsge_diff_normalization_sql
+from .ocsge_occupation_du_sol_normalization_sql import (
+    ocsge_occupation_du_sol_normalization_sql,
+)
+from .ocsge_zone_construite_normalization_sql import (
+    ocsge_zone_construite_normalization_sql,
+)
+
+__all__ = [
+    "ocsge_artif_normalization_sql",
+    "ocsge_diff_normalization_sql",
+    "ocsge_occupation_du_sol_normalization_sql",
+    "ocsge_zone_construite_normalization_sql",
+    "get_normalized_fields",
+    "ocsge_artif_diff_normalization_sql",
+]
